Multicluster support for kiali cache
